Splitted board methods btw files
