+++ src/board/mod.rs
@@ -1,0 +1,3 @@
+pub mod board;
+pub mod fen_handling;
+pub mod mask_handling;
