+++ src/debug/mod.rs
@@ -1,3 +1,2 @@
 pub mod divide;
-pub mod knight_movement;
 pub mod tree;
